+++ frontend-react/e2e/pages/about/about-side-navigation.ts
