+++ frontend-react/e2e/pages/about/about.ts
@@ -1,4 +1,4 @@
-import { BasePage, BasePageTestArgs } from "./BasePage";
+import { BasePage, BasePageTestArgs } from "../BasePage";
 
 export class AboutPage extends BasePage {
     constructor(testArgs: BasePageTestArgs) {
