+++ frontend-react/e2e/pages/about/roadmap.ts
@@ -1,17 +1,8 @@
-import { BasePage, BasePageTestArgs } from "../BasePage";
+import { Page } from "@playwright/test";
 
-export class AboutRoadmapPage extends BasePage {
-    constructor(testArgs: BasePageTestArgs) {
-        super(
-            {
-                url: "/about/roadmap",
-                title: "ReportStream roadmap",
-                heading: testArgs.page.getByRole("heading", {
-                    name: "Roadmap",
-                    exact: true,
-                }),
-            },
-            testArgs,
-        );
-    }
+export const URL_ROADMAP = "/about/roadmap";
+export async function goto(page: Page) {
+    await page.goto(URL_ROADMAP, {
+        waitUntil: "domcontentloaded",
+    });
 }
