--- frontend-react/e2e/pages/roadmap.ts
@@ -1,8 +1,0 @@
-import { Page } from "@playwright/test";
-
-export const URL_ROADMAP = "/about/roadmap";
-export async function goto(page: Page) {
-    await page.goto(URL_ROADMAP, {
-        waitUntil: "domcontentloaded",
-    });
-}
