+++ frontend-react/e2e/spec/all/about/about-page.spec.ts
