+++ frontend-react/e2e/spec/all/about/our-network-page.spec.ts
@@ -1,23 +1,19 @@
 import { expect, test } from "@playwright/test";
 
-import * as internalLinks from "../../helpers/internal-links";
-import * as sideNav from "../../pages/about-side-navigation";
-import * as roadmap from "../../pages/roadmap";
-import { URL_ROADMAP } from "../../pages/roadmap";
-
+import * as sideNav from "../../../pages/about/about-side-navigation";
+import * as ourNetwork from "../../../pages/about/our-network";
 test.describe(
-    "Product roadmap page",
+    "Our network page",
     {
         tag: "@smoke",
     },
     () => {
         test.beforeEach(async ({ page }) => {
-            await roadmap.goto(page);
+            await ourNetwork.goto(page);
         });
 
         test("has correct title", async ({ page }) => {
-            await expect(page).toHaveURL(URL_ROADMAP);
-            await expect(page).toHaveTitle(/Product roadmap/);
+            await expect(page).toHaveTitle(/Our network/);
         });
 
         test.describe("Side navigation", () => {
@@ -49,23 +45,6 @@ test.describe(
             test("has Release notes link", async ({ page }) => {
                 await sideNav.clickReleaseNotes(page);
                 await expect(page).toHaveURL(/.*about\/release-notes/);
-            });
-        });
-
-        test.describe("Additional resources Links", () => {
-            test("has News", async ({ page }) => {
-                await internalLinks.clickOnInternalLink("div", "CardGroup", "News", page);
-                await expect(page).toHaveURL(/.*about\/news/);
-            });
-
-            test("has Release notes", async ({ page }) => {
-                await internalLinks.clickOnInternalLink("div", "CardGroup", "Release notes", page);
-                await expect(page).toHaveURL(/.*about\/release-notes/);
-            });
-
-            test("has Developer resources", async ({ page }) => {
-                await internalLinks.clickOnInternalLink("div", "CardGroup", "Developer resources", page);
-                await expect(page).toHaveURL(/.*developer-resources/);
             });
         });
     },
